cmd/check: add topology conflit check
